% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addt.R
\name{addt}
\alias{addt}
\title{Produces the t-test for an additional explanatory variable}
\usage{
addt(y, x, w, intercept = TRUE, la, nocheck = FALSE, trace)
}
\arguments{
\item{y}{A vector with \code{n} elements that contains the response variable.}

\item{x}{An \code{n x p} data matrix (\code{n} observations and \code{p}
 variables) of explanatory variables
 (also called 'regressors') of dimension n x (p-1) where p denotes the number of
 explanatory variables including the intercept.

 Rows of \code{X} represent observations, and columns represent variables. By default,
 there is a constant term in the model, unless you explicitly remove it using input
 option intercept, so do not include a column of 1s in \code{X}.
 Missing values (NAs)
 and infinite values (Inf's) are allowed, since observations (rows) with missing or
 infinite values will automatically be excluded from the computations.}

\item{w}{the added variable, a vector containing the additional explanatory variable
whose \code{t-test} must be computed.}

\item{intercept}{Indicator for the constant term (intercept) in the fit, defaults to \code{intercept=TRUE}.}

\item{la}{transformation parameter. It specifies for which Box Cox transformation parameter
is necessary to compute the \code{t} statistic for the additional variable. If \code{la} is missing
(default) no transformation is used. For example \code{la=0.5} tests square root transformation}

\item{nocheck}{Wheather to check the input arguments. If \code{nocheck=TRUE} no check is performed on
matrix \code{y} and matrix \code{X}. Notice that \code{y} and \code{X} are left unchanged. In other words
the additional column of ones for the intercept is not added. By default \code{nocheck=FALSE}.}

\item{trace}{Whether to print intermediate results. Default is \code{trace=FALSE}.}
}
\value{
An object of class \code{addt}, will be returned which is basically a list
 containing two matrices, \code{Un} and \code{BB}:
 \enumerate{
 \item \code{b} estimate of the slope for additional explanatory variable
 \item \code{S2add} estimate of \eqn{s^2} of the model which contains the additional explanatory variable
 \item \code{Tadd} t-statistic for additional explanatory variable
 \item \code{pval} p-value of the t-statistic
 }
}
\description{
Produces the t-test for an additional explanatory variable.
}
\examples{

   data(wool)
   XX <- wool
   y <- log(XX[, ncol(XX)])
   w <- XX[, ncol(XX)-1]
   X <- XX[, 1:(ncol(XX)-2), drop=FALSE]
   ## Intercept=TRUE (default)
   ##  - has effect only when nocheck=FALSE (default)
   (out <- addt(y, X, w, intercept=TRUE, nocheck=FALSE))

}
\references{
Atkinson A.C. and Riani M. (200), \emph{Robust Diagnostic Regression Analysis} Springer Verlag, New York.
}
\author{
FSDA team, \email{valentin.todorov@chello.at}
}
