% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{diabetes}
\alias{diabetes}
\title{Diabetes data}
\format{
A data frame with 145 rows and 4 variables
The variables are as follows:

 \itemize{
     \item glucose: Area under plasma glucose curve after a three hour oral glucose tolerance test (OGTT).
     \item insulin: Area under plasma insulin curve after a three hour oral glucose tolerance test (OGTT).
     \item sspg: Steady state plasma glucose.
     \item class: The type of diabete: \code{Normal}, \code{Overt}, and \code{Chemical}.
 }
}
\source{
Reaven, G. M. and Miller, R. G. (1979). An attempt to define the nature of
 chemical diabetes using a multidimensional analysis.
 \emph{Diabetologia} 16:17-24.
}
\usage{
data(diabetes)
}
\description{
The diabetes dataset, introduced by Reaven and Miller (1979),
 consists of 145 observations (patients). For each patient three
 measurements are reported: plasma glucose response to oral glucose,
 plasma insulin response to oral glucose, degree of insulin resistance.
}
\examples{
 data(diabetes)
 head(diabetes)

}
\keyword{datasets}
