% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{emilia2001}
\alias{emilia2001}
\title{Demographic data from the 341 miniciplaities in Emilia Romagna (an Italian region).}
\format{
A data frame with 341 rows and 28 variables
The variables are as follows:

\itemize{
  \item less10: population aged less than 10
  \item more75: population aged more than 75
  \item single single-member families
  \item divorced": divorsed
  \item widows: widows and widowers
  \item graduates: population aged more than 25 who are graduates
  \item no_education:  of those aged over 6 having no education
  \item employed: activity rate
  \item unemployed: unemployment rate
  \item increase_popul: standardised natural increase in population
  \item migration: standardised change in population due to migration
  \item birth_92_94: average birth rate over 1992-94
  \item fecundity: three-year average birth rate amongst women of child-bearing age
  \item houses: occupied houses built since 1982
  \item houses_2WCs: occupied houses with 2 or more WCs
  \item houses_heating: occupied houses with fixed heating system
  \item TV: TV licence holders
  \item cars: number of cars for 100 inhabitants
  \item luxury_cars: luxury cars
  \item hotels: working in hotels and restaurants
  \item banking: working in banking and finance
  \item income: average declared income amongst those filing income tax returns
  \item income_tax_returns: inhabitants filing income tax returns
  \item factories: residents employed in factories and public services
  \item factories_more10: employees employed in factories withy more tha 10 employees
  \item factories_more50: employees employed in factories withy more tha 50 employees
  \item artisanal: artisanal enterprises
  \item entrepreneurs: enterpreneous and skilled self-employed among those of working age
}

 @references
 Atkinson, A. C., Riani, M., and Cerioli, A. (2004). \emph{Exploring Multivariate Data with the Forward Search}. Springer-Verlag, New York.
}
\usage{
data(emilia2001)
}
\description{
A data set containing 28 demographic variables for 341 municipalities in Emilia Romagna (an Italian region).
}
\keyword{datasets}
