% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{fishery}
\alias{fishery}
\title{Fishery data.}
\format{
A data frame with 677 rows and 2 variables
}
\usage{
data(fishery)
}
\description{
The fishery data consist of 677 transactions of a fishery product in Europe.
 For each transaction the Value in 1000 euro and the quantity in Tons are reported.
 Data extracted from monthly aggregates (flows) of trade declarations (Riani et al. 2008).
 The dataset is formed by 677 flows of a fishery product imported in the European Union from
 a third country in a period of one year. Among the many variables available we provide:
 \enumerate{
 \item \code{x}: the quantity of the trade flow;
 \item \code{y}: the value of the trade flow;
 }
 By regressing the variable \code{value} against the \code{quantity} one can see that the
 dataset is characterized by the presence of a mixture of linear groups, which roughly
 correspond to the clusters indicated by the subject matter expert. Riani et al. (2008)
 have shown how the FS can estimate such a mixture, allocate the units to the components
 of the mixture and identify in the dataset possible outliers, i.e. units that do not
 belong to any component. The three identified components are consistent with the clusters
 identified by the subject matter experts. The dataset is one among thousands of similar
 datasets that have to be analyzed automatically, for which there is no subject matter
 classification available.
}
\keyword{datasets}
