% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FSMenvmmd.R
\name{FSMenvmmd}
\alias{FSMenvmmd}
\title{Computes the theoretical envelopes of Minimum MD outside subset during the search}
\usage{
FSMenvmmd(n, v, init, prob, scaled = FALSE)
}
\arguments{
\item{n}{Number of observations.}

\item{v}{Number of variables.}

\item{init}{Search initialization. It specifies the point where to initialize the search and start
monitoring required diagnostics. If it is not specified by default it will be equal to \code{floor(n * 0.6)}.}

\item{prob}{row vector of \code{k} quantiles for which envelopes have to be computed. 
The default is to produce 1 per cent, 50 per cent and 99 per cent envelopes.}

\item{scaled}{It indicates how to compute the envelopes. Boolean.
If \code{scaled=TRUE} the envelopes are produced for
scaled Mahalanobis distances (no consistency factor is applied) else the traditional 
consistency factor is applied (this is the default)}
}
\value{
An object of class \code{FSMenvmmd}, will be returned which is basically a list
 containing the matrix \code{MMDenv} that is a \code{n-m0+1 x length(prob)+1} array containing the envelopes
              for the requested quantiles.
 \enumerate{
            \item 1st col = fwd search index from \code{m0 to n-1};
            \item 2nd col = envelope for quantile \code{prob[1]};
            \item 3rd col = envelope for quantile \code{prob[2]};
            \item ...;
            \item \code{(k+1)} col = envelope for quantile \code{prob[k]}.  
}
}
\description{
Computes the theoretical envelopes of Minimum MD outside subset during the search.
}
\examples{

library(fsdac)

out<-FSMenvmmd(n=100, v=3, prob=c(0.01, 0.05, 0.5, 0.95, 0.99), init=9, scaled=FALSE)
plot(out)

}
\references{
Riani, M., Atkinson, A.C. and Cerioli, A. (2009), Finding an unknown
  number of multivariate outliers, "Journal of the Royal Statistical
  Society Series B", Vol. 71, pp. 201-221.
}
\author{
FSDA team, \email{valentin.todorov@chello.at}
}
