% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FSMinvmmd.R
\name{FSMinvmmd}
\alias{FSMinvmmd}
\title{Converts values of minimum Mahalanobis distance into confidence levels}
\usage{
FSMinvmmd(mmd, v, n)
}
\arguments{
\item{mmd}{Matrix of distances of dimension \code{n-m0}.
1st col = forward search index
2nd col = minimum Mahalanobis distance.}

\item{v}{Number of variables.}
}
\value{
An object of class \code{FSMinvmmd}, will be returned which is basically a list
 containing the matrix \code{MMDinvMat} that is a \code{n-init+1 x 3} array containing
 information about confidence levels plotted in normal coordinates.
 \enumerate{
            \item 1st col = fwd search index from \code{init to n-1};
            \item 2nd col = confidence level of each value of \code{mmd};
            \item 3rd col = confidence level in normal coordinates.
            \item 50\% per cent conf level becomes \code{norminv(0.50)=0};
            \item 99 per cent conf level becomes \code{norminv(0.99)=2.33}.  
}
}
\description{
Computes the theoretical envelopes of Minimum MD outside subset during the search.
}
\examples{

library(fsdac)

out <- FSMenvmmd(n=100, v=3, prob=c(0.05, 0.5, 0.95), init=9, scaled=FALSE)
(FSMinvmmd(mmd=out, v=3, n=100))

}
\references{
Atkinson, A.C. and Riani, M. (2006), Distribution theory and
simulations for tests of outliers in regression, "Journal of
Computational and Graphical Statistics", Vol. 15, pp. 460-476.

Riani, M. and Atkinson, A.C. (2007), Fast calibrations of the forward
search for testing multiple outliers in regression, "Advances in Data
Analysis and Classification", Vol. 1, pp. 123-141.
}
\author{
FSDA team, \email{valentin.todorov@chello.at}
}
