% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FSRenvmdr.R
\name{FSRenvmdr}
\alias{FSRenvmdr}
\title{Computes the theoretical envelopes of Minimum Deletion Residual outside subset during the search}
\usage{
FSRenvmdr(n, v, init, prob)
}
\arguments{
\item{n}{Number of observations.}

\item{v}{Number of variables.}

\item{init}{Search initialization. It specifies the point where to initialize the search and start
monitoring required diagnostics. If it is not specified by default it will be equal to \code{floor(n * 0.6)}.}

\item{prob}{row vector of \code{k} quantiles for which envelopes have to be computed. 
The default is to produce 1 per cent, 50 per cent and 99 per cent envelopes.}
}
\value{
An object of class \code{FSRenvmdr}, will be returned which is basically a list
 containing the matrix \code{MDRenv} that is a \code{n-init+1 x length(prob)+1} array containing the envelopes
              for the requested quantiles.
 \enumerate{
            \item 1st col = fwd search index from \code{init to n-1};
            \item 2nd col = envelope for quantile \code{prob[1]};
            \item 3rd col = envelope for quantile \code{prob[2]};
            \item ...;
            \item \code{(k+1)} col = envelope for quantile \code{prob[k]}.  
}
}
\description{
Computes the theoretical envelopes of Minimum Deletion Residual outside subset during the search.
}
\examples{

library(fsdac)

out<-FSRenvmdr(n=100, v=3, prob=c(0.01, 0.05, 0.5, 0.95, 0.99), init=9)
plot(out)

}
\references{
Atkinson, A.C. and Riani, M. (2006), Distribution theory and
simulations for tests of outliers in regression, "Journal of
Computational and Graphical Statistics", Vol. 15, pp. 460-476.
Riani, M. and Atkinson, A.C. (2007), Fast calibrations of the forward
search for testing multiple outliers in regression, "Advances in Data
Analysis and Classification", Vol. 1, pp. 123-141.
}
\author{
FSDA team, \email{valentin.todorov@chello.at}
}
