% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{geyser2}
\alias{geyser2}
\title{Old Faithful Geyser Data.}
\format{
A data frame with 271 rows and 2 variables
The variables are as follows:

\itemize{
  \item Eruption length: The eruption length in minutes.
  \item Previous eruption length: The length of the previous eruption in minutes.
}
}
\usage{
data(geyser2)
}
\description{
A bivariate data set obtained from the Old Faithful Geyser, containing the eruption
 length and the length of the previous eruption for 271 eruptions of this geyser in minutes.
}
\references{
Garcia-Escudero, L.A., Gordaliza, A. (1999). Robustness properties of k-means and trimmed k-means,
     \emph{Journal of the American Statistical Assoc.}, Vol.\strong{94}, No.447, 956-969.

     Haerdle, W. (1991). \emph{Smoothing Techniques with Implementation in S}, New York: Springer.
}
\keyword{datasets}
