% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{hawkins}
\alias{hawkins}
\title{Hawkins data.}
\format{
A data frame with 128 rows and 9 variables
}
\usage{
data(hawkins)
}
\description{
These data, simulated by Hawkins, consist of 128 observations
     and eight explanatory variables \code{(X1, ..., X8)} and one dependent variable, \code{y}.
}
\keyword{datasets}
