% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lxs.R
\name{LXS}
\alias{LXS}
\title{Computes the Least Median of Squares (LMS) or Least Trimmed Squares (LTS) estimators}
\usage{
LXS(
  y,
  x,
  intercept = TRUE,
  lms = 1,
  rew = FALSE,
  bonflevoutX,
  alpha = 0.5,
  h,
  conflev = 0.975,
  nsamp,
  nomes = FALSE,
  msg = TRUE,
  nocheck = FALSE,
  csave = FALSE,
  trace = FALSE
)
}
\arguments{
\item{y}{A vector with \code{n} elements that contains the response variable.}

\item{x}{An \code{n x p} data matrix (\code{n} observations and \code{p}
 variables) of explanatory variables
 (also called 'regressors') of dimension n x (p-1) where p denotes the number of
 explanatory variables including the intercept.

 Rows of \code{X} represent observations, and columns represent variables. By default,
 there is a constant term in the model, unless you explicitly remove it using input
 option intercept, so do not include a column of 1s in \code{X}.
 Missing values (NaN's)
 and infinite values (Inf's) are allowed, since observations (rows) with missing or
 infinite values will automatically be excluded from the computations.}

\item{intercept}{Indicator for the constant term (intercept) in the fit, defaults to \code{intercept=TRUE}.}

\item{lms}{estimation method, If \code{lms=1} (default) Least Median of Squares is computed,
 else if \code{lms=2} or  \code{lms='lts'} fast LTS with all default options is used else
 if \code{lms} is a scalar different from 1 and 2 standard lts is used (without concentration steps) else
 if \code{lms} is a list or a named vector fast lts (with concentration steps) is used. In this case the user
 can control the following options:
 \enumerate{
 \item \code{refsteps} - number of refining iterations on each subsample (default is \code{refsteps=3}).
     If \code{refsteps=0}, this  means 'raw-subsampling' without iterations.
 \item \code{reftol} - default value for the tolerance used for the refining steps (default is \code{reftol=1e-6}).
 \item \code{bestr} - number of 'best betas' to remember from the subsamples. These will be
     later iterated until convergence (default is \code{bestr=5}).
 \item \code{refstepsbestr} - number of refining iterations for each best subset (default is \code{refstepsbestr=50}).
 \item \code{reftolbestr} - default value for the tolerance for the refining steps for each of
     the best subsets (default is \code{reftolbestr=1e-8}).
}}

\item{rew}{wheather to reweight the LTS/LMS estomates. If \code{rew=TRUE} the reweighted version of LTS/LMS is used and
the output quantities refer to the reweighted version, else no reweighting is performed (default).}

\item{bonflevoutX}{control outlier detection in the designmatrix. If the design matrix \code{X} contains
several high leverage units (that is units which are very far from the bulk of the data), it may happen
that the best subset may include some of these units. If \code{boflevoutX} is not missing, outlier
detection procedure \code{FSM()} is applied to the design matrix \code{X}, using name/pair option
\code{bonflev=bonflevoutX}. The extracted subsets which contain at least one unit declared as outlier
in the \code{X} space by \code{FSM()} are removed (more precisely they are treated as singular subsets)
from the list of candidate subsets to find the \code{LXS()} solution. By default (\code{bonflevoutX} is missing)
\code{FSM()} is not invoked.}

\item{alpha}{the percentage (roughly) of squared residuals whose sum will be minimized,
by default \code{alpha=0.5}. In general, alpha must be between 0.5 and 1.}

\item{h}{The number of observations that have determined the least trimmed squares
estimator,  scalar. \code{h} is an integer greater or equal than \code{p} but smaller
then \code{n}. Generally \code{h=[0.5*(n+p+1)]} (default value).}

\item{conflev}{confidence level which is used to declare outliers. It could be \code{95, 0.975, 0.99}
(individual alpha) or \code{1-0.05/n, 1-0.025/n, 1-0.01/n} (simultaneous alpha). The default is \code{conflev=0.975}.}

\item{nsamp}{number of subsamples which will be extracted to find the robust estimator. If \code{nsamp=0}
 all subsets will be extracted. They will be \code{n choose p}.

 Remark: if the number of all possible subset is <1000 the default is to extract all subsets
 otherwise just 1000.}

\item{nomes}{controls whether to display or not on the screen messages about estimated
time to compute LMS (LTS). If \code{nomes=TRUE} no message about estimated time to compute
LMS (LTS) is displayed, else if \code{nomes=FALSE} (default), a message about
estimated time is displayed.}

\item{msg}{Level of output to display. It controls whether to display or not messages on the screen.
 If \code{msg=TRUE} (default) messages are displayed on the screen about step of the fwd search else
no message is displayed on the screen.}

\item{nocheck}{Wheather to check the input arguments. If \code{nocheck=TRUE} no check is performed on
matrix \code{y} and matrix \code{X}. Notice that \code{y} and \code{X} are left unchanged. In other words
the additional column of ones for the intercept is not added. By default \code{nocheck=FALSE}.}

\item{csave}{wheather to return the optional matrix \code{C} containing the indexes of
the subsamples extracted for computing the estimate (the so called elemental sets).}

\item{trace}{Whether to print intermediate results. Default is \code{trace=FALSE}.}
}
\value{
An object of class \code{lts} or \code{lms} will be returned which is basically a list
 containing the following elements:
 \enumerate{
 \item \code{rew}: wheather reweighting was applied. If \code{rew=TRUE}, all subsequent output refers to
     reweighted estimates.
 \item \code{coefficients}: LTS (LMS) coefficient estimates, including the intercept when \code{intercept=TRUE}.
 \item \code{bs}: a vector containing the units forming the subset associated with bLMS (bLTS).
 \item \code{residuals}: a vector containing the standardized residuals from the regression.
 \item \code{scale}: scale estimate of the residuals.
 \item \code{weights}: a vector containing weights. The elements of this vector are 0 or 1.
     These weights identify the \code{h} observations which are used to compute the final LTS (LMS) estimate.
     \code{sum(weights)=h} if there is not a perfect fit, otherwise \code{sum(weights)} can be greater than \code{h}.
 \item \code{h}: the number of observations that have determined the LTS (LMS) estimator, i.e. the value of h.
 \item \code{outliers}: a vector containing the list of the units declared as outliers
     using confidence level specified in \code{conflev}.
 \item \code{conflev}: confidence level which is used to declare outliers.
 \item \code{singsub}: number of subsets wihtout full rank. Notice that if this number is greater than
     \code{0.1*(number of subsamples)} a warning is produced on the screen.
 \item \code{y}: the response variable.
 \item \code{X}: the predictor matrix.
 \item \code{C}: the matrix containing the indexes of the subsamples extracted for computing
     the estimate (the so called elemental sets) (only of \code{csave=TRUE}).

 }
}
\description{
LXS computes the Least Median of Squares (LMS) or Least Trimmed Squares (LTS) estimators.
}
\examples{

 data(heart, package="robustbase")

 ## Default method works with 'x'-matrix and y-var and all default optional arguments
 heart.x <- data.matrix(heart[, 1:2]) # the X-variables
 heart.y <- heart[,"clength"]
 (out <- LXS(heart.y, heart.x))

 data(stackloss)
 LXS(stackloss$stack.loss, stackloss[, -4])

}
\references{
Rousseeuw P.J. and Leroy A.M. (1987), \emph{Robust regression and outlier detection}, Wiley.
}
\author{
FSDA team, \email{valentin.todorov@chello.at}
}
