% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{M5data}
\alias{M5data}
\title{Mixture M5 Data.}
\format{
A data frame with 2000 rows and 3 variables
The first two columns are the two variables. The last column is the true classification vector where symbol "0" stands for the contaminating data points.
}
\source{
Garcia-Escudero, L.A., Gordaliza, A., Matran, C. and Mayo-Iscar, A. (2008). A General Trimming Approach to Robust Cluster Analysis,
     \emph{Annals of Statistics}, Vol.\strong{36}, 1324-1345. \doi{10.1214/07-AOS515}.
}
\usage{
data(M5data)
}
\description{
A bivariate data set obtained from three normal bivariate distributions with different scales and
 proportions 1:2:2. One of the components is strongly overlapping with another one. A 10%% background
 noise is added uniformly distributed in a rectangle containing the three normal components and not
 strongly overlapping with the three mixture components. A precise description of the M5 data set
 can be found in Garcia-Escudero et al. (2008).
}
\keyword{datasets}
