% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{multiple_regression}
\alias{multiple_regression}
\title{Multiple regression data showing the effect of masking (Atkinson and Riani, 2000).}
\format{
A data frame with 60 rows and 4 variables
The variables are as follows:

\itemize{
  \item X1
  \item X2
  \item X3
  \item y
}

 @references
 Atkinson, A. C., and Riani, M. (2000). \emph{Robust Diagnostic Regression Analysis}. Springer-Verlag, New York.
}
\usage{
data(multiple_regression)
}
\description{
There are 60 observations on a response y with the values of three explanatory variables.
 The scatter plot matrix of the data shows y increasing with each of x1, x2 and x3.
 The plot of residuals against fitted values shows no obvious pattern. However the
 FS finds that there are 6 masked outliers.
}
\keyword{datasets}
