% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{mussels}
\alias{mussels}
\title{Mussels data.}
\format{
A data frame with 82 rows and 5 variables
}
\usage{
data(mussels)
}
\description{
These data, introduced by Cook and Weisberg (1994), consist of 82 observations on horse mussels from
 New Zeland. The variables are shell length, width, height, mass and muscle mass
}
\keyword{datasets}
