% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{poison}
\alias{poison}
\title{Poison}
\format{
A data frame with 48 rows and 7 variables: six explanatory and one response variable.
}
\source{
G. E. P. Box and D. R. Cox (1964). An Analysis of Transformations,
 \emph{ournal of the Royal Statistical Society. Series B}, \bold{26}2 pp. 211--252.
}
\usage{
data(poison)
}
\description{
The poison data (by Box and Cox, 1964) are about the time to death of animals in a \code{3 x 4}
 factorial experiment with four observations at each factor combination. There are no outliers
 or influential observations that cannot be reconciled with the greater part of the data by a
 suitable transformation.
}
\examples{
 data(poison)
 head(poison)

}
\keyword{datasets}
