% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sreg.R
\name{Sreg}
\alias{Sreg}
\title{Computes S estimators in linear regression}
\usage{
Sreg(
  y,
  x,
  intercept = TRUE,
  bdp = 0.5,
  rhofunc = c("bisquare", "optimal", "hyperbolic", "hampel", "mdpd"),
  rhofuncparam,
  nsamp = 1000,
  refsteps = 3,
  reftol = 1e-06,
  refstepsbestr = 50,
  reftolbestr = 1e-08,
  minsctol = 1e-07,
  bestr = 5,
  conflev = 0.975,
  msg = TRUE,
  nocheck = FALSE,
  csave = FALSE,
  trace = FALSE
)
}
\arguments{
\item{y}{A vector with \code{n} elements that contains the response variable.}

\item{x}{An \code{n x p} data matrix (\code{n} observations and \code{p}
 variables) of explanatory variables
 (also called 'regressors') of dimension n x (p-1) where p denotes the number of
 explanatory variables including the intercept.

 Rows of \code{x} represent observations, and columns represent variables. By default,
 there is a constant term in the model, unless you explicitly remove it using input
 option intercept, so do not include a column of 1s in \code{x}.
 Missing values (NaN's)
 and infinite values (Inf's) are allowed, since observations (rows) with missing or
 infinite values will automatically be excluded from the computations.}

\item{intercept}{Indicator for the constant term (intercept) in the fit, defaults to \code{intercept=TRUE}.}

\item{bdp}{breakdown point -  measures the fraction of outliers the algorithm
should resist. Any value greater than 0 but smaller or equal than 0.5 will do.
Note that for given bdp nominal efficiency is determined automatically.}

\item{rhofunc}{'rho' function. A string which specifies the rho function which will
be used to weight the residuals. Possible values are: \code{c('bisquare', 'optimal', 'hyperbolic', 'hampel', 'mdpd')}.
The choice 'bisquare' uses Tukey's \eqn{\rho} and \eqn{\psi} functions,
'optimal' uses the optimal \eqn{\rho} and \eqn{\psi} functions;
'hyperbolic' uses hyperbolic \eqn{\rho} and \eqn{\psi} functions,
'hampel' uses Hampel \eqn{\rho} and \eqn{\psi} functions;
'mdpd' uses Minimum Density Power Divergence \eqn{\rho} and \eqn{\psi} functions.
The default is \code{rhofunc='bisquare'}.}

\item{nsamp}{number of subsamples which will be extracted to find the robust estimator.
If \code{nsamp=0} all subsets will be extracted. They will be \code{(n choose p)}.
Remark: if the number of all possible subset is <1000 the default is to extract all subsets
otherwise just 1000.}

\item{refsteps}{number of refining iterations in each subsample (default is \code{refsteps=3}).
\code{refsteps=0} means "raw-subsampling" without iterations.}

\item{reftol}{tolerance for the refining steps. The default value is \code{reftol=1e-6}.}

\item{refstepsbestr}{number of refining iterations for each best subset. (default is \code{refstepsbestr=50}).}

\item{reftolbestr}{tolerance for the refining steps for each of the best subsets. The default value is \code{reftolbestr=1e-8}.}

\item{minsctol}{tolerance for the iterative procedure for finding the minimum
value of the scale for each subset and each of the best subsets. The default value is \code{minsctol=1e-7}.}

\item{bestr}{number of 'best betas' to remember from the subsamples. These will be
later iterated until convergence. Default is \code{bestr=5}.}

\item{conflev}{confidence level which is used to declare units as outliers. Usually
\code{conflev=c(0.95, 0.975, 0.99)} (individual alpha) or
\code{conflev=c(1-0.05/n, 1-0.025/n, 1-0.01/n)} (simultaneous alpha).
Default is \code{conflev=0.975}.}

\item{msg}{Level of output to display. It controls whether to display or not messages on the screen.
 If \code{msg=TRUE} (default) messages are displayed on the screen about step of the fwd search else
no message is displayed on the screen.}

\item{nocheck}{Wheather to check the input arguments. If \code{nocheck=TRUE} no check is performed on
matrix \code{y} and matrix \code{x}. Notice that \code{y} and \code{x} are left unchanged. In other words
the additional column of ones for the intercept is not added. By default \code{nocheck=FALSE}.}

\item{trace}{Whether to print intermediate results. Default is \code{trace=FALSE}.}

\item{rhofuncparam:}{additional parameters for the specified rho function (one or more double numbers).
For the hyperbolic rho function it is possible to set the value of \code{k = sup CVC}
(the default value is \code{k=4.5}).
For the Hampel rho function it is possible to define the parameters \code{a}, \code{b} and \code{c}
(the default values are \code{a=2, b=4, c=8} or \code{rhofuncparam=c(2, 4, 8)}).}
}
\value{
An object of class \code{Sreg} will be returned which is basically a list
 containing the following elements:
 \enumerate{
 \item \code{coefficients}: S estimates of the regression coefficients, including the intercept when \code{intercept=TRUE}.
 \item \code{residuals}: a vector containing the standardized residuals from the regression.
 \item \code{fitted.values}: a vector containing the fitted values.
 \item \code{weights}: a vector containing the estimated weights.
 \item \code{scale}: scale estimate of the residuals.
 \item \code{outliers}: a vector containing the list of the units declared as outliers
     using the confidence level specified in \code{conflev}.
 \item \code{bs}: a \code{p x 1} vector containing the units forming best subset
     associated with the S estimate of the regression coefficients.
 \item \code{conflev} confidence level which is used to declare outliers.
     Remark: the returned \code{conflev} will be used to draw the horizontal line
     (confidence band) in the plot.
 \item \code{singsub}: number of subsets wihtout full rank. Notice that
     \code{singsub > 0.1*(number of subsamples)} will produce a warning.
 \item \code{rhofunc}: a character string identifying the \eqn{\rho} function which has been used.
 \item \code{rhofuncparam}: a vector with additional parameters for the specified \eqn{\rho}
     function which have been used. For the hyperbolic \eqn{\rho} function this is the value
     of \code{k=sup CVC}. For the Hampel \eqn{\rho} function these are the parameters \code{a, b and c}.
     This field is present only if the input argument \code{rhofunc} is  \code{'hyperbolic'} or \code{'hampel'}.
 \item \code{singularity}: if present, this element indicatres a singularity condition.
 \item \code{y}: the response variable.
 \item \code{X}: the predictor matrix.
 }
}
\description{
Sreg computes S estimators in linear regression.
}
\examples{

 data(hbk, package="robustbase")
 XX <- hbk
 y <- XX[, ncol(XX)]
 X <- XX[, 1:(ncol(XX)-1), drop=FALSE]

 (out <- Sreg(y, X))

}
\references{
Maronna, R.A., Martin D. and Yohai V.J. (2006), \emph{Robust Statistics, Theory and Methods},
     Wiley, New York.
}
\author{
FSDA team, \email{valentin@todorov.at}
}
