% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unibiv.R
\name{unibiv}
\alias{unibiv}
\title{Detects univariate and bivariate outliers}
\usage{
unibiv(Y, madcoef = 1.4815, rf = 0.95, robscale = 1, trace = FALSE)
}
\arguments{
\item{Y}{Input data. A \code{n x v} data matrix; \code{n} observations and \code{v} variables.
Rows of Y represent observations, and columns represent variables. Missing values (NaN's) and
infinite values (Inf's) are allowed, since observations (rows) with missing or infinite values
will automatically be excluded from the computations.}

\item{madcoef}{scaled MAD. Scalar. Coefficient which is used to scale MAD
coefficient to have a robust estimate of dispersion.  The
default is \code{1.4815} so that \code{1.4815*MAD(N(0,1))=1}.
Remark: if \code{mad=median(y-median(y))=0} then the interquartile
range is used. If also the interquartile range is 0
than the \code{MD} (mean absolute deviation) is used.  In
other words \code{MD=mean(abs(y-mean(Y)))}.}

\item{robscale}{how to compute dispersion. Specifies the
                 statistical indexes to use to compute the dispersion of
                 each variable and the correlation among each pair of variables.
\enumerate{
  \item robscale=1 (default): the program uses the median correlation
                 and the MAD as estimate of the dispersion of each variable;
  \item robscale=2: the correlation coefficient among ranks is used
                 (Spearman's rho) and the MAD as estimate of the dispersion
                 of each variable;
  \item robscale=3: the correlation coefficient is based on Kendall's tau b
                 and the MAD as estimate of the dispersion of each variable;
  \item  robscale=4: tetracoric correlation coefficient is used and the MAD
                 as estimate of the dispersion of each variable;
                 otherwise the correlation and the dispersion of the variables are
                 computed using the traditional (non robust) formulae
                 around the univariate medians.
 }}
}
\value{
\item \code{fre}   \code{n x 4} matrix which contains details about the univariate and
          bivariate outliers.
 \itemize {
  \item        1st col = index of the units;
  \item        2nd col = number of times unit has been declared
               univariate outliers;
  \item        3rd col = number of times unit has been declared
               bivariate outlier;
  \item         4th col = pseudo \code{MD} as sum of bivariate \code{MD}.
 }
}
\description{
unibiv has the purpose of detecting univariate and bivariate outliers.
}
\examples{
library(fsdac)
data(swissbanknotes)
Y = swissbanknotes
Y = Y[1:200,1:6]

(out=unibiv(Y))


}
\references{
Riani, M., Zani S. (1997). An iterative method for the detection of
 multivariate outliers, \emph{Metron}, Vol. LV, pp. 101-117.
}
\author{
FSDA team, \email{valentin@todorov.at}
}
