% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{wool}
\alias{wool}
\title{Wool data (Box and Cox, 1964)}
\format{
A data frame with 27 rows and 4 variables
The variables are as follows:

\itemize{
  \item length
  \item amplitude
  \item load
  \item cycles
}

 @references
 Atkinson, A. C., and Riani, M. (2000). \emph{Robust Diagnostic Regression Analysis}. Springer-Verlag, New York.
}
\usage{
data(wool)
}
\description{
Number of cycles to failure of samples of worsted yarn in a 33 experiment (Box and Cox, 1964).
 The wool data give the number of cycles to failure of a worsted yarn under cycles of repeated
 loading. The results are from a single 33 factorial experiment. The three factors and their
 levels are:
 \itemize{
 - x1: length of test specimen (25, 30, 35 cm)
 - x2: amplitude of loading cycle (8, 9, 10 mm)
 - x3: load (40, 45, 50 g).
 }
 The number of cycles to failure ranges from 90, for the shortest specimen subject to the most
 severe conditions, to 3,636 for observation 19 which comes from the longest specimen subjected
 to the mildest conditions. In their analysis Box and Cox(1964) recommend that the data be fitted
 after the log transformation of y. The FS plots explain the effect of the ordering of the data
 during the FS on the estimates of regression coefficients and the error variance and on
 a score statistic for transformation of the response.
}
\keyword{datasets}
