% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{X}
\alias{X}
\title{Simulated data X.}
\format{
A data frame with 200 rows and 2 variables
}
\usage{
data(X)
}
\description{
The X dataset has been simulated by Gordaliza, Garcia-Escudero and Mayo-Iscar during the Workshop
 ADVANCES IN ROBUST DATA ANALYSIS AND CLUSTERING held in Ispra on October 21st-25th 2013. It is
 a bivariate dataset of 200 observations. It presents two parallel components without contamination.
}
\keyword{datasets}
